+++ backend/canisters/user_index/impl/src/jobs/distribute_airdrop_neurons.rs
@@ -1,17 +1,29 @@
 use crate::model::initial_airdrop_queue::InitialAirdropEntry;
 use crate::{mutate_state, RuntimeState};
 use candid::Principal;
+<<<<<<< HEAD
 use canister_client::make_c2c_call;
+=======
+>>>>>>> master
 use ic_cdk::api::call::{CallResult, RejectionCode};
 use ic_cdk_timers::TimerId;
 use ic_sns_governance::pb::v1::manage_neuron::{AddNeuronPermissions, Command, RemoveNeuronPermissions, Split};
 use ic_sns_governance::pb::v1::manage_neuron_response::Command as CommandResponse;
+<<<<<<< HEAD
 use ic_sns_governance::pb::v1::{ManageNeuron, ManageNeuronResponse, NeuronPermissionList};
 use std::cell::Cell;
 use std::time::Duration;
 use tracing::trace;
 use types::{CanisterId, SnsNeuronId, UserId};
 use utils::hasher::hash_bytes;
+=======
+use ic_sns_governance::pb::v1::{ManageNeuron, NeuronPermissionList};
+use rand_core::RngCore;
+use std::cell::Cell;
+use std::time::Duration;
+use tracing::{error, trace};
+use types::{CanisterId, SnsNeuronId, UserId};
+>>>>>>> master
 
 const ALL_PERMISSIONS: [i32; 11] = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10];
 
@@ -44,6 +56,10 @@ fn try_get_next(state: &mut RuntimeState) -> Option<AirdropNeuronArgs> {
         user_id: e.user_id,
         neuron_controller: e.neuron_controller,
         neuron_stake_e8s: e.neuron_stake_e8s,
+<<<<<<< HEAD
+=======
+        memo: state.env.rng().next_u64(),
+>>>>>>> master
         this_canister_id: state.env.canister_id(),
         governance_canister_id: state.data.openchat_governance_canister_id,
         source_neuron_id: state.data.initial_airdrop_neuron_id.unwrap(),
@@ -62,7 +78,12 @@ async fn process_next(args: AirdropNeuronArgs) {
         neuron_controller: args.neuron_controller,
         neuron_stake_e8s: args.neuron_stake_e8s,
     };
+<<<<<<< HEAD
     if airdrop_neuron_to_user(args).await.is_err() {
+=======
+    if let Err(error) = airdrop_neuron_to_user(args).await {
+        error!(?error, args = ?entry, "Failed to distribute airdrop neuron");
+>>>>>>> master
         mutate_state(|state| state.data.initial_airdrop_queue.mark_failed(entry));
     }
 }
@@ -70,9 +91,15 @@ async fn process_next(args: AirdropNeuronArgs) {
 async fn airdrop_neuron_to_user(args: AirdropNeuronArgs) -> CallResult<SnsNeuronId> {
     let neuron_id = split(
         args.governance_canister_id,
+<<<<<<< HEAD
         args.neuron_controller,
         args.source_neuron_id,
         args.neuron_stake_e8s,
+=======
+        args.source_neuron_id,
+        args.neuron_stake_e8s,
+        args.memo,
+>>>>>>> master
     )
     .await?;
 
@@ -84,12 +111,19 @@ async fn airdrop_neuron_to_user(args: AirdropNeuronArgs) -> CallResult<SnsNeuron
 
 async fn split(
     governance_canister_id: CanisterId,
+<<<<<<< HEAD
     recipient_principal: Principal,
     source_neuron_id: SnsNeuronId,
     stake_e8s: u64,
 ) -> CallResult<[u8; 32]> {
     let memo = u64::from_be_bytes(hash_bytes(recipient_principal.as_slice())[..8].try_into().unwrap());
 
+=======
+    source_neuron_id: SnsNeuronId,
+    stake_e8s: u64,
+    memo: u64,
+) -> CallResult<[u8; 32]> {
+>>>>>>> master
     let args = ManageNeuron {
         subaccount: source_neuron_id.to_vec(),
         command: Some(Command::Split(Split {
@@ -98,11 +132,15 @@ async fn split(
         })),
     };
 
+<<<<<<< HEAD
     let response: ManageNeuronResponse =
         make_c2c_call(governance_canister_id, "manage_neuron", args, candid::encode_one, |r| {
             candid::decode_one(r)
         })
         .await?;
+=======
+    let response = sns_governance_canister_c2c_client::manage_neuron(governance_canister_id, &args).await?;
+>>>>>>> master
 
     match response.command.unwrap() {
         CommandResponse::Split(s) => Ok(s.created_neuron_id.unwrap().id.try_into().unwrap()),
@@ -122,11 +160,15 @@ async fn add_all_permissions(governance_canister_id: CanisterId, neuron_id: [u8;
         })),
     };
 
+<<<<<<< HEAD
     let response: ManageNeuronResponse =
         make_c2c_call(governance_canister_id, "manage_neuron", args, candid::encode_one, |r| {
             candid::decode_one(r)
         })
         .await?;
+=======
+    let response = sns_governance_canister_c2c_client::manage_neuron(governance_canister_id, &args).await?;
+>>>>>>> master
 
     match response.command.unwrap() {
         CommandResponse::AddNeuronPermission(_) => Ok(()),
@@ -150,11 +192,15 @@ async fn remove_all_permissions(
         })),
     };
 
+<<<<<<< HEAD
     let response: ManageNeuronResponse =
         make_c2c_call(governance_canister_id, "manage_neuron", args, candid::encode_one, |r| {
             candid::decode_one(r)
         })
         .await?;
+=======
+    let response = sns_governance_canister_c2c_client::manage_neuron(governance_canister_id, &args).await?;
+>>>>>>> master
 
     match response.command.unwrap() {
         CommandResponse::RemoveNeuronPermission(_) => Ok(()),
@@ -167,6 +213,10 @@ struct AirdropNeuronArgs {
     user_id: UserId,
     neuron_controller: Principal,
     neuron_stake_e8s: u64,
+<<<<<<< HEAD
+=======
+    memo: u64,
+>>>>>>> master
     this_canister_id: CanisterId,
     governance_canister_id: CanisterId,
     source_neuron_id: SnsNeuronId,