+++ backend/external_canisters/sns_governance/api/src/updates/manage_neuron.rs
@@ -1,0 +1,4 @@
+use ic_sns_governance::pb::v1::{ManageNeuron, ManageNeuronResponse};
+
+pub type Args = ManageNeuron;
+pub type Response = ManageNeuronResponse;